Enable IRQ before handling the event triggered by fw
Sometimes the time of INT triggered by fw would be too shorter than
then time that TP driver couldn't take. Enabling IRQ before dealing with
data from fw may have a risk that the kthread could be overlap after the
next INT event is coming. Howevent, this is a way of workaround to solve the issue
of coming time, I'll take more time to test it.