Fix the issue of thread could block in sleep state
Some of platform, particularly in kernel 3-.19, may cause the problem
showing the message as below :

'do not call blocking ops when !TASK_RUNNING; state=1'

We use a new function called wait_woken() for a wakeup. The key is in
the use of a special wakeup function called woken_wait_function to deal
with what happens when the wakeup arrives.

More information see https://lwn.net/Articles/627930/